requirements.txt
Flask==2.0.3
mysql-connector-python-rf==2.2.2
requests==2.27.1
uWSGI==2.0.20
python-dotenv==0.20.0
